+++ HISTORY.md
@@ -6,6 +6,10 @@ TODOS/WISHES
 1. add SERIAL execution as method wait_all_piped! paired up with wait_all_parallel() 
 
 
+- BIG REF! for thread_instance
+- use QThread (terminate + ...)
+- with to add meta cumulative funks
+
 0.0.5 (2023-12-19)
 -------------------
 - fix nothread